Sync 1.2.x branch with upstream

+++ src/gnutls/globals.h
@@ -6,7 +6,7 @@
  * This is free software; see Copyright file in the source
  * distribution for preciese wording.
  *
- * Copyright (C) 2002-2016 Aleksey Sanin <aleksey@aleksey.com>. All Rights Reserved.
+ * Copyright (C) 2002-2024 Aleksey Sanin <aleksey@aleksey.com>. All Rights Reserved.
  */
 #ifndef __XMLSEC_GLOBALS_H__
 #define __XMLSEC_GLOBALS_H__
@@ -40,9 +40,9 @@
                     (const char*)(errorObject),             \
                     (errorFunction),                        \
                     XMLSEC_ERRORS_R_CRYPTO_FAILED,          \
-                    "gcrypt error: %ld: %s: %s",            \
-                    (long)(errCode),                        \
-                    xmlSecErrorsSafeString(source),         \
+                    "gcrypt error: %u: %s: %s",   \
+                    (errCode),                              \
+                    xmlSecErrorsSafeString(source),      \
                     xmlSecErrorsSafeString(message)         \
         );                                                  \
     }
@@ -62,9 +62,9 @@
                     (const char*)(errorObject),             \
                     (errorFunction),                        \
                     XMLSEC_ERRORS_R_CRYPTO_FAILED,          \
-                    "gnutls error: %ld: %s",                \
-                    (long)(errCode),                        \
-                    xmlSecErrorsSafeString(message)         \
+                   "gnutls error: %d: %s",             \
+                    (errCode),                              \
+                    xmlSecErrorsSafeString(message)     \
         );                                                  \
     }
 
@@ -85,9 +85,9 @@
                     (const char*)(errorObject),             \
                     (errorFunction),                        \
                     XMLSEC_ERRORS_R_CRYPTO_FAILED,          \
-                    msg  "gnutls error: %ld: %s",           \
+                    msg  "gnutls error: %d: %s",            \
                     (param),                                \
-                    (long)(errCode),                        \
+                    (errCode),                              \
                     xmlSecErrorsSafeString(message)         \
         );                                                  \
     }
